third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-sizes-js-innerhtml.html

<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script>
    if (window.testRunner)
        testRunner.dumpAsText();

    addEventListener("DOMContentLoaded", function() {
        if (window.testRunner && sessionStorage.pageReloaded)
            testRunner.dumpResourceRequestCallbacks();
        window.testdiv = document.createElement("div");
        window.testdiv.innerHTML = '<img id="foo" src="resources/blue-100-px-square.png" srcset="../../hidpi/resources/image-set-1x.png 200w, ../../hidpi/resources/image-set-2x.png 400w, resources/image-set-4x.png 800w">';
    }, false);
    addEventListener("load", function() {
        shouldBe('window.testdiv.children[0].currentSrc', '"resources/image-set-4x.png"');
    }, false);
</script>
</head>

<body id="body">
    <img id="foo">
</body>
</html>
