third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-sizes-js-innerhtml-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS window.testdiv.children[0].currentSrc is "resources/image-set-4x.png"
